refactor: according mentor's req
