+++ .github/pull_request_template.md
@@ -4,4 +4,3 @@
 3. Deploy:
 4. Done ..2023 / deadline ..2023
 5. Score: 100 / 100
-
