+++ README.md
@@ -72,5 +72,3 @@ Prettier
 ## Authors
 
 - [@vikkitorry](https://github.com/vikkitorry)
-
-
