+++ src/app/services/types/serviceTypes.ts
@@ -1,16 +1,3 @@
-export const ERROR_MESSAGE = 'There is nothing here';
-
-export const BASE_URL = 'https://api.myshows.me/v2/rpc/';
-export const defaultPageSize = 30;
-export const lang = 'en';
-export const id = 1;
-export const jsonrpc = '2.0';
-export enum METHOD {
-  getAllShows = 'shows.Get',
-  getShow = 'shows.GetById',
-  post = 'POST',
-}
-
 export type Response = {
   jsonrpc: string;
   result: ShowSchema[];
