+++ src/components/Card/Detailed/Detailed.module.scss
