+++ src/components/ErrorBoundary/ErrorBoundary.tsx
@@ -1,5 +1,5 @@
 import React, { ErrorInfo, ReactNode, Suspense } from 'react';
-import { ErrorPage } from '../../../pages/error/ErrorPage';
+import { ErrorPage } from '../../pages/error/ErrorPage';
 
 interface ErrorBoundaryProps {
   children: ReactNode;
