+++ src/components/SearchBar/SearchBar.tsx
@@ -1,12 +1,12 @@
 import { Button, ButtonSize } from '../Button/Button';
 import cls from './SearchBar.module.scss';
 import Input from '../Input/Input';
-import React, { useState, useCallback, memo } from 'react';
+import React, { useState, useCallback } from 'react';
 import { SEARCH_LOCALSTORAGE_KEY } from '../../utils/constants/Constants';
-import { MainPageRoutes } from '../../app/providers/router/routeConfig/routeConfig';
+import { MainPageRoutes } from '../../app/router/routeConfig/routeConfig';
 import { SetURLSearchParams } from 'react-router-dom';
 import { DropDown } from '../DropDown/DropDown';
-import { defaultPageSize } from '../../app/providers/services/types/serviceTypes';
+import { defaultPageSize } from '../../app/services/variables/variables';
 
 const NUM_OF_ITEMS_VALUES = [defaultPageSize.toString(), '20', '10', '5'];
 
@@ -16,7 +16,7 @@ interface ISearchBarProps {
   setNumOfItems: (value: number) => void;
 }
 
-export const SearchBar = memo((props: ISearchBarProps) => {
+export const SearchBar = (props: ISearchBarProps) => {
   const { setSearch, setSearchParams, setNumOfItems } = props;
   const [inputValue, setInputValue] = useState<string | undefined>(
     localStorage.getItem(SEARCH_LOCALSTORAGE_KEY) || undefined
@@ -26,22 +26,25 @@ export const SearchBar = memo((props: ISearchBarProps) => {
     setInputValue(value);
   }, []);
 
-  const changeQuery = () => {
+  const changeQuery = useCallback(() => {
     setSearchParams((searchParams) => {
       searchParams.set(MainPageRoutes.PAGE, '1');
       return searchParams;
     });
-  };
+  }, [setSearchParams]);
 
   const onSubmit = () => {
     changeQuery();
     setSearch(inputValue ? inputValue : '');
   };
 
-  const onSelectNumOfItems = (value: string) => {
-    changeQuery();
-    setNumOfItems(+value);
-  };
+  const onSelectNumOfItems = useCallback(
+    (value: string) => {
+      changeQuery();
+      setNumOfItems(+value);
+    },
+    [changeQuery, setNumOfItems]
+  );
 
   return (
     <div className={cls.SearchBar}>
@@ -58,4 +61,4 @@ export const SearchBar = memo((props: ISearchBarProps) => {
       <DropDown values={NUM_OF_ITEMS_VALUES} onChange={onSelectNumOfItems} />
     </div>
   );
-});
+};
