+++ src/pages/main/MainPage.tsx
@@ -2,7 +2,7 @@ import cls from './mainPage.module.scss';
 import CardsHandler from '../../components/Cards/CardsHandler';
 import { useState, useEffect, useCallback } from 'react';
 import { useSearchParams } from 'react-router-dom';
-import { MainPageRoutes } from '../../app/providers/router/routeConfig/routeConfig';
+import { MainPageRoutes } from '../../app/router/routeConfig/routeConfig';
 import { CardsHandlerSize } from '../../components/Cards/CardsHandler';
 import { Outlet } from 'react-router-dom';
 
@@ -12,7 +12,7 @@ export const MainPage = () => {
   const isShowOpen = searchParams.get(MainPageRoutes.SHOW);
 
   useEffect(() => {
-    setIsDetailedOpen(!!isShowOpen);
+    setIsDetailedOpen(Boolean(isShowOpen));
   }, [isShowOpen]);
 
   const closeDetailed = useCallback(() => {
