+++ src/pages/noResults/NoResultsPage.module.scss
@@ -16,7 +16,7 @@
     content: '';
     width: 100%;
     height: 100%;
-    background: url(./assets/Rick.jpg) no-repeat;
+    background: url(../../assets/Rick.jpg) no-repeat;
     background-size: cover;
     border-radius: 50%;
   }
